Move integration tests outside the lib.
